fix(72): adjust Reading queries parameter interactions
Previously, Reading queries would be ANDed together.
For example, if you'd given it both region and address params, it would
search for readings in the addresses AND the regions. Which is stupid,
given that addresses have regions, and if the address is outside the
regions provided to the query, no readings would be returned.
So now, they are ORed together;
Given a set of regions and addresses, reading on either regions OR
addresses provided will be returned.